requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
distlib==0.3.8
Django==5.0.8
django-celery-beat==2.6.0
django-environ>=0.11.2
django-timezone-field==7.0
django-filter>=24.3
djangorestframework>=3.15.2
filelock==3.15.4
kombu==5.4.0
markdown==3.6
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2==2.9.9
PyMySQL==1.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests>=2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
vine==5.1.0
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
